tools/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
docker==7.1.0
idna==3.7
PyYAML==6.0.1
requests==2.32.3
robotframework==7.0.1
urllib3==2.2.2
